Add PriorityLevel support for various Cosmos operations
The program adds support for setting the PriorityLevel of Cosmos operations. This change allow different Cosmos operations to have different priorities. The PriorityLevel feature is only enabled when the PREVIEW compilation symbol is defined, in which case ItemRequestOptions and QueryRequestOptions arguments include a PriorityLevel property. This property is used to set the priority of Cosmos operations.