Refactor build workflow and remove unnecessary packages
This commit involves significant changes to the build process which is now triggered on every push. This ensures that all code changes are accounted for and tested, regardless of whether they're part of a pull request. In addition, the build now takes place only on ubuntu-latest OS, which simplifies the process and reduces resource consumption. Moreover, the 'Nerdbank.GitVersioning' packages have been removed from both 'Atc.Cosmos.csproj' and 'Directory.Build.props' files, as they were found to be unnecessary. The build process now also packages the library and uploads the produced artifacts. Changes also include a VERSION variable that is set to 1.0 concatenated with the Github run number.