Update post-integration.yml for more robust testing and artifact handling
In this commit, several changes have been made to the `post-integration.yml` file to enhance our testing and artifact management. Firstly, in the 'run unit tests' command, `-p:DefineConstants=PREVIEW` was appended, to ensure that tests run with the proper configuration.

Secondly, the `continue-on-error: true` property was added to the SonarCloud install scanner step. This allows the process to continue even if an error occurs in this step, ensuring subsequent steps are still executed.

An unnecessary step of creating library package for pre-release was removed and "push packages" step was modified to push all .nupkg files, resulting in simpler flow.

Lastly, an 'upload build artifacts' step was added. This uses the `actions/upload-artifact@v3` action to upload any .nupkg files and the README.md file as build artifacts. This will aid in storing and retrieving build products for each workflow run.