Enable workflow to continue after package push errors
An update was made to the post-integration GitHub action workflow. It now allows the workflow to proceed even if errors were encountered while pushing packages to the GitHub Package Registry. This change was implemented to avoid the entire process stopping due to minor errors during the package push stage. Now, we can ensure the upload of build artifacts will not be blocked due to an unsuccessful package push.