Add additional interfaces to README.md
In this commit, some new interfaces for handling Cosmos document resources were added to the README.md file. This has been done to inform users about the expanded functionality of low-priority reading and writing that is now available in the preview version of the package. This should help users better leverage the package's capabilities.